More concurrency changes for the sqlalchemy store.
This one makes sure that revisions do not collide.

It does this by using unique ids for revisions. This has the
the benefit of making joins cheaper as well. The con is that
now revisions for a tiddler will continue to be sequences, but
with gaps: 1, 5, 8, 500, 2574 are the 5 revisions of one tiddler.

In the long run this is a good thing, but may confuse people at first.