Correct editing existing policies.
In the previous code if there existed a read policy with user X
and we wanted to add user Y, the result would be a read policy of
user Y, but no X.

The _store_policy code is almost certainly wrong, but works. It
needs significant review to determine where it can be streamlined.
Much of what it is doing sqlalchemy should be taking care of for us.