Go back to eagerly loading bag and recipe policies.
This had been made lazy in early changes to avoid loading policies
when listing bags and recipes. Then, later, listing bags and recipes
stopped loading them from the store at all, so that optimization then
became a cost for the later loading of individual bags and recipes, when
we will definitely want the policy.