make processing of lists of bags and recipes lazy
The tiddlyspace.com mysql-slow log shows that /bags and /recipes
results in frequent slow requests. We've established that slow
request prevent the system from effectively using resources.

Also, the objects returned from list_recipes and list_bags are
empty anyway, so the data is not used, so getting it is redundant.