Tune up _load_tiddler to deal with rare situations.
Sometimes a revision will have no associated entry in the text
table. When this happens the previous _load_tiddler will fail
rather than just setting tiddler.text to '', which is what should
happen.

This same problem doesn't happen with fields and tags because the
orm treats those as empty lists.

The exception handling in this situation was out of date for
the calling code. It used to be that an IndexError could occur
if a revision didn't exist, but checking for the revision happens
in the calling code not the _load_tiddler method, so we no longer
need to trap an IndexError and turn it into a NoTiddlerError,
that happens above, where a NoResult becomes a NoTiddler.