Do not load lists of recipes or bags lazily from db.
We _always_ need the bag or recipe policy when listing them so
it makes no sense to hold off on getting the policy information.
For a small number of bags or recipes the hold off is not costly,
but when the number of entities gets over several thousand it means
that a request for a list of bags results in several thousand SQL
calls. This brings it down to one.

The entire codebase for this is a mishmash of balancing acts like this
one.