Establish on delete cascades in schema.
This only has impact on database types that support them (such as
innodb), but where it does it means that:

   delete from bag where name='foobar'

will delete all the tiddlers, revisions, tags, etc "in" that bag.

This is needed for raw SQL access to things like the tiddlyspace
database.

However: Because the same code needs to support different types of
databases, sqlalchemy still implements its own cascade. This can
be controlled by setting passive_deletes to True, but it's not
clear if this can be meaningful per table type. More research
required.