Do raw INSERTs into the text table.
This allows us to set a 'DELAYED' prefix when mysql is
being used with myisam tables. This means that the mysql thread
can return before the insert is complete. This is good because the
write of the big binary tiddler's text field could block for
quite a long time, either waiting on table locks or simply
waiting on disk.

This does introduce a slight risk that the content won't get
written but that's the price we have to pay if we want to
write big files and not ruin interactivity for other users.