Increase support for referential integrity on tiddler deletes.
Fix #3

Tiddlyspace.com revealed that text table was not being deleted
on cascades. Turns out that was because it was myisam and the
other tables were innodb. Now that all the tables are innodb
it is fine.

However the investigation revealed that there were two different
styles of deletes going on in the code: session delete() and
query delete(). The latter does a raw delete sql query and
allows foreign key cascades to do their work. The former gets
everything in the session and then deletes each one by one.

Everything was switched to the latter and the sqlite engine was
made to support referential integrity.