An unlimited UnicodeText is more generic.
That is: it is more likely to work with more rdbms systems.

Notably, it does not work with mysql, resulting in a TEXT, which is
too short, there a special mapping to LONGTEXT is required. However
that mapping is required whether a length is given to UnicodeText
or not, so this change does not impact that situation.